Fixed plotting

+++ snudda/utils/export_connection_matrix.py
@@ -50,13 +50,13 @@ class SnuddaExportConnectionMatrix(object):
                 sparse_data[idx, :] = [x, y, self.con_mat[x, y]]
 
             # np.savetxt(self.out_file, sparse_data, delimiter=",", fmt="%d")
-            np.save(self.out_file, sparse_data)
+            np.save(self.out_file, sparse_data.astype(np.int32))
 
             # Test to verify
             for row in sparse_data:
                 assert self.con_mat[row[0], row[1]] == row[2]
         else:
-            np.save(self.out_file, self.con_mat)
+            np.save(self.out_file, self.con_mat.astype(np.int32))
             # np.savetxt(self.out_file, self.con_mat, delimiter=",", fmt="%d")
 
         print("Writing " + self.out_file_meta)
@@ -128,7 +128,7 @@ class SnuddaExportConnectionMatrix(object):
         save_matrix[np.isnan(save_matrix)] = np.inf
 
         # np.savetxt(delay_file, save_matrix, delimiter=",", fmt="%d")
-        np.save(delay_file, save_matrix)
+        np.save(delay_file, save_matrix.astype(np.float32))
 
     ############################################################################
 
@@ -142,7 +142,7 @@ class SnuddaExportConnectionMatrix(object):
         if plot:
             self.plot_matrix(dist_matrix)
 
-        np.save(dist_file, dist_matrix)
+        np.save(dist_file, dist_matrix.astype(np.float32))
 
         # np.savetxt(dist_file, dist_matrix, delimiter=",", fmt="%d")
